startup-advisor-mcp/pyproject.toml
[project]
name = "03-startup-advisor-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.2",
    "strands-agents-tools>=0.1.1",
    "termcolor>=3.1.0",
]
